moved InventoryTest file into correct filepath and directory
