+++ src/test/java/coffeemaker/domain/InventoryTest.java
@@ -1,4 +1,4 @@
-package coffeemaker;
+package coffeemaker.domain;
 
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,8 +8,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import coffeemaker.exceptions.InventoryException;
-
-import coffeemaker.domain.*;
 
 public class InventoryTest {
     /**
